+++ src/core/server/commands/getVehicle.ts
@@ -1,6 +1,7 @@
 import * as alt from 'alt-server';
 import { addCommand } from '../systems/chat';
 import { getDescription } from '../../shared/commands/commandList';
+import { getVectorInFrontOfPlayer } from '../utility/vector';
 
 addCommand('getvehicle', handleCommand);
 
@@ -15,8 +16,10 @@ function handleCommand(player: alt.Player, name: string): void {
         return;
     }
 
+    const forwardPosition = getVectorInFrontOfPlayer(player, 5);
+
     try {
-        const veh = new alt.Vehicle(name, player.pos.x, player.pos.y, player.pos.z, 0, 0, 0);
+        const veh = new alt.Vehicle(name, forwardPosition.x, forwardPosition.y, forwardPosition.z, 0, 0, 0);
     } catch (err) {
         player.send(`Vehicle model is not valid`);
     }